+++ src/output/p_avc.h
@@ -19,8 +19,9 @@
 
 class avc_video_packetizer_c: public generic_video_packetizer_c {
 protected:
-  int m_nalu_size_len_src, m_nalu_size_len_dst;
-  int64_t m_max_nalu_size;
+  int m_nalu_size_len_src{}, m_nalu_size_len_dst{};
+  int64_t m_max_nalu_size{}, m_track_default_duration{-1};
+  debugging_option_c m_debug_fix_bistream_timing_info;
 
 public:
   avc_video_packetizer_c(generic_reader_c *p_reader, track_info_c &p_ti, double fps, int width, int height);
@@ -37,5 +38,5 @@ protected:
   virtual void extract_aspect_ratio();
   virtual void setup_nalu_size_len_change();
   virtual void change_nalu_size_len(packet_cptr packet);
-  virtual void remove_filler_nalus(memory_c &data) const;
+  virtual void process_nalus(memory_c &data) const;
 };